+++ nvim/lua/plugin_manager.lua
@@ -130,4 +130,18 @@ require("lazy").setup({
 			"nvim-lua/plenary.nvim",
 		},
 	},
+
+	-- 這個外掛可以在 neovim 上刷 leetcode
+	{
+		"kawre/leetcode.nvim",
+		build = ":TSUpdate html",
+		dependencies = {
+			"nvim-telescope/telescope.nvim",
+			"nvim-lua/plenary.nvim",
+			"MunifTanjim/nui.nvim",
+			"nvim-treesitter/nvim-treesitter",
+			"rcarriga/nvim-notify",
+			"nvim-tree/nvim-web-devicons",
+		},
+	},
 })